feat: Refactor UI structure and enhance search result display
