+++ config.py
@@ -1,3 +1,5 @@
+from dotenv import load_dotenv
+load_dotenv()
 import os
 import json
 import gspread
@@ -31,40 +33,41 @@ RIGHT_COLUMN_SEARCH_START_RATIO = 0.65
 RIGHT_COLUMN_SEARCH_WIDTH_RATIO = 0.20
 
 # --- Botが投稿するEmbedの画像URL ---
+AUTHOR_NAME = "ファインモーション"
 AUTHOR_ICON_URL = "https://cdn.discordapp.com/attachments/1407605158161940480/1407617349355442197/2-removebg-preview.png"
 SEARCH_THUMBNAIL_URL = "https://cdn.discordapp.com/attachments/1407605158161940480/1407605207222718574/-removebg-preview.png"
 RANKING_THUMBNAIL_URL = "https://cdn.discordapp.com/attachments/1407605158161940480/1407626444397482086/IMG_3186-removebg-preview.png"
 MYBOX_THUMBNAIL_URL = "https://cdn.discordapp.com/attachments/1407605158161940480/1407627785463136301/IMG_3187-removebg-preview.png"
 REGISTER_THUMBNAIL_URL = "https://cdn.discordapp.com/attachments/1407605158161940480/1407625108683755620/-removebg-preview.png"
 
+
 # --- Googleスプレッドシートのキー ---
 SPREADSHEET_KEY = "1NxsYfkptjaFGeVMQh9-5WtcaXpgf0qcvsLyRMvo5anw"
 
+# --- Google Credentials (Universal Setup) ---
+google_creds_json_str = os.getenv('GOOGLE_CREDENTIALS_JSON')
+gc = None 
 
-# ===============================================================
-# ▼▼▼ Google認証情報の読み込み処理（Render最終版） ▼▼▼
-# ===============================================================
-gc = None
-credentials_json_str = os.getenv('GOOGLE_CREDENTIALS_JSON')
-
-if credentials_json_str:
+if google_creds_json_str:
     try:
-        # 1. Secretの中身を一時的なファイルに書き出す
-        creds_filename = "temp_google_creds.json"
-        with open(creds_filename, "w") as f:
-            f.write(credentials_json_str)
-
-        # 2. 古いgspreadが唯一理解できる「ファイル名で読み込む」方法で認証する
-        scope = [
-            'https://www.googleapis.com/auth/spreadsheets',
-            'https://www.googleapis.com/auth/drive'
-        ]
-        creds = ServiceAccountCredentials.from_json_keyfile_name(creds_filename, scope)
-        gc = gspread.authorize(creds)
+        credentials_dict = json.loads(google_creds_json_str)
         
-        print("✅ Google認証情報、正常に読み込み完了。")
+        # 全てのGoogleライブラリが参照する、一時的な認証情報ファイルを作成
+        temp_creds_path = "temp_google_creds.json"
+        with open(temp_creds_path, "w") as f:
+            json.dump(credentials_dict, f)
+        
+        # 環境変数に「認証情報ファイルの場所」をセット
+        os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = temp_creds_path
+        
+        # gspreadクライアントもこのファイルを使って認証
+        gc = gspread.service_account(filename=temp_creds_path)
+        
+        print("✅ Google認証情報、ファイル経由で正常に設定完了。")
 
+    except json.JSONDecodeError:
+        print("❌ エラー: .envのGOOGLE_CREDENTIALS_JSONが正しいJSON形式ではありません。")
     except Exception as e:
         print(f"❌ Google認証情報の処理中にエラー: {e}")
 else:
-    print("❌ エラー: Google認証情報が環境変数に設定されていません。")
+    print("❌ エラー: .envにGOOGLE_CREDENTIALS_JSONが設定されていません。")
